delta/drop v1, promote v2 to the sole delta API
Remove the local-build Delta v1 surface and the unreleased DeltaV2 namespace;
the v2 server-built implementation now backs sdk.delta.* and the bare top-level
exports. Replace getBridgeInfo with getBridgeRoutes, and point getQuote at
/v2/quote returning the v2 route-based DeltaPrice (overloads/modes unchanged).
Migrate examples to the v2 flow.
