+++ src/index.ts
@@ -138,13 +138,13 @@ import type {
   PaginatedResponse,
 } from './types';
 
+// ── Delta ─────────────────────────────────────────────────────────────────
+// Delta is v2 (server-built orders). The previous local-build v1 surface was
+// removed; every Delta symbol below refers to the v2 implementation.
+
+// Shared on-chain order-struct types
 import type {
   DeltaAuctionOrder,
-  DeltaAuction,
-  DeltaAuctionStatus,
-  DeltaAuctionTransaction,
-  BridgeMetadata,
-  BridgeStatus,
   Bridge,
   ExternalDeltaOrder,
   ProductiveDeltaOrder,
@@ -158,21 +158,31 @@ import type {
   DeltaAmountsSellSlippage,
   DeltaAmountsBuySlippage,
   DeltaAmountsExplicit,
-  DeltaAuctionDelta,
-  DeltaAuctionFillable,
-  DeltaAuctionTWAP,
-  DeltaAuctionTWAPBuy,
-  DeltaAuctionExternal,
-  DeltaAuctionProductive,
   DeltaOrderUnion,
-  DeltaAuctionUnion,
   UnifiedDeltaOrderData,
 } from './methods/delta/helpers/types';
+// v2 auction, price & route types
+import type {
+  BuiltDeltaOrder,
+  DeltaAuction,
+  DeltaOrderStatus,
+  DeltaTokenSide,
+  DeltaTransaction,
+  DeltaPrice,
+  DeltaPriceToken,
+  DeltaTokenAmount,
+  BridgeTag,
+  DeltaRoute,
+  DeltaRouteStep,
+  DeltaRouteBridge,
+  DeltaRouteBridgeContractParams,
+  BridgeRoute,
+} from './methods/delta/types';
+
 import {
-  BuildDeltaOrderDataParams,
   BuildDeltaOrderFunctions,
+  BuildDeltaOrderParams,
   constructBuildDeltaOrder,
-  SignableDeltaOrderData,
 } from './methods/delta/buildDeltaOrder';
 import {
   constructPostDeltaOrder,
@@ -183,7 +193,7 @@ import {
 import {
   constructSignDeltaOrder,
   SignDeltaOrderFunctions,
-} from './methods/delta/signDeltaOrder';
+} from './methods/delta';
 import {
   constructPreSignDeltaOrder,
   PreSignDeltaOrderFunctions,
@@ -195,16 +205,11 @@ import {
 import {
   constructGetDeltaPrice,
   GetDeltaPriceFunctions,
-  DeltaPrice,
-  BridgePrice,
-  AvailableBridge,
   DeltaPriceParams,
 } from './methods/delta/getDeltaPrice';
 import {
   constructGetDeltaOrders,
   GetDeltaOrdersFunctions,
-  DeltaOrderFilterByStatus,
-  DeltaOrderFromAPI,
 } from './methods/delta/getDeltaOrders';
 import {
   ApproveTokenForDeltaFunctions,
@@ -215,15 +220,19 @@ import {
   GetPartnerFeeFunctions,
 } from './methods/delta/getPartnerFee';
 import {
-  constructGetBridgeInfo,
-  GetBridgeInfoFunctions,
-  BridgeInfo,
+  constructGetBridgeRoutes,
+  GetBridgeRoutesFunctions,
   BridgeProtocolResponse,
-} from './methods/delta/getBridgeInfo';
+} from './methods/delta/getBridgeRoutes';
 import {
   constructIsTokenSupportedInDelta,
   IsTokenSupportedInDeltaFunctions,
 } from './methods/delta/isTokenSupportedInDelta';
+import {
+  constructGetAgentsList,
+  GetAgentsListFunctions,
+  AgentList,
+} from './methods/delta/getAgentsList';
 
 import {
   constructBuildExternalDeltaOrder,
@@ -231,10 +240,6 @@ import {
   BuildExternalDeltaOrderParams,
 } from './methods/delta/buildExternalDeltaOrder';
 import type { SignableExternalOrderData } from './methods/delta/helpers/buildExternalOrderData';
-import {
-  constructSignExternalDeltaOrder,
-  SignExternalDeltaOrderFunctions,
-} from './methods/delta/signExternalDeltaOrder';
 import {
   constructPostExternalDeltaOrder,
   PostExternalDeltaOrderFunctions,
@@ -247,16 +252,12 @@ import {
 
 import {
   BuildTWAPDeltaOrderParams,
-  BuildTWAPSellOrderParams,
-  BuildTWAPBuyOrderParams,
+  BuildTWAPSellDeltaOrderParams,
+  BuildTWAPBuyDeltaOrderParams,
   BuildTWAPDeltaOrderFunctions,
   constructBuildTWAPDeltaOrder,
 } from './methods/delta/buildTWAPDeltaOrder';
 import type { SignableTWAPOrderData } from './methods/delta/helpers/buildTWAPOrderData';
-import {
-  constructSignTWAPDeltaOrder,
-  SignTWAPDeltaOrderFunctions,
-} from './methods/delta/signTWAPDeltaOrder';
 import {
   constructPostTWAPDeltaOrder,
   PostTWAPDeltaOrderFunctions,
@@ -278,6 +279,9 @@ import {
 } from './methods/quote/getQuote';
 import {
   CancelDeltaOrderFunctions,
+  CancelDeltaOrder,
+  SignCancelDeltaOrderRequest,
+  PostCancelDeltaOrderRequest,
   constructCancelDeltaOrder,
 } from './methods/delta/cancelDeltaOrder';
 import {
@@ -291,6 +295,7 @@ import {
   CancelDeltaOrderData,
   SignableCancelDeltaOrderData,
 } from './methods/delta/helpers/buildCancelDeltaOrderData';
+import { SignableDeltaOrderData } from './methods/delta/helpers/buildDeltaOrderData';
 
 export { constructSwapSDK, SwapSDKMethods } from './methods/swap';
 
@@ -310,32 +315,6 @@ export {
   SubmitExternalDeltaOrderParams,
   SubmitTWAPDeltaOrderParams,
 } from './methods/delta';
-
-// Delta v2 is exposed as a single namespace so it can ship alongside v1 without
-// colliding at the top level. v2 source uses unsuffixed names internally
-// (constructBuildDeltaOrder, BuiltDeltaOrder, DeltaPrice, ...); consumers reach
-// them via `DeltaV2.constructBuildDeltaOrder` etc.
-//
-// Migration plan when deprecating v1 (breaking change):
-//   1. Replace the bare `export { ... } from './methods/delta'` block above with
-//      `export * from './methods/deltaV2'` (or named re-exports of v2 symbols).
-//      Top-level `constructPostDeltaOrder` now resolves to v2.
-//   2. Move v1 behind its own namespace for backcompat:
-//      `export * as DeltaV1 from './methods/delta'`.
-//   3. Keep `export * as DeltaV2 from './methods/deltaV2'` as a redundant alias
-//      so code written against the v2 namespace today keeps compiling.
-//   4. Mirror this on the bundled SDKs in sdk/full.ts and sdk/simple.ts:
-//      `sdk.delta.*` flips to v2 (use `constructAllDeltaOrdersHandlers` from
-//      './methods/deltaV2'), and add `sdk.deltaV1.*` for backcompat.
-//
-// Migration plan when dropping v1 entirely:
-//   - Delete the `DeltaV1` namespace export and `sdk.deltaV1` field.
-//   - Delete `src/methods/delta/`.
-//   - Delete the `DeltaV2` re-export alias (consumers can move to bare names).
-//   - The `methods/deltaV2/` folder can be renamed to `methods/delta/` (and the
-//     `/delta/v2/` URL prefix in the leaf modules updated if the server has by
-//     then collapsed v1 and v2 endpoints).
-export * as DeltaV2 from './methods/deltaV2';
 
 export type {
   TransactionParams,
@@ -407,21 +386,20 @@ export {
   constructCancelDeltaOrder,
   constructDeltaTokenModule,
   constructApproveTokenForDelta,
+  constructGetAgentsList,
   // External Delta methods
   constructBuildExternalDeltaOrder,
-  constructSignExternalDeltaOrder,
   constructPostExternalDeltaOrder,
   constructPreSignExternalDeltaOrder,
   // TWAP Delta methods
   constructBuildTWAPDeltaOrder,
-  constructSignTWAPDeltaOrder,
   constructPostTWAPDeltaOrder,
   constructPreSignTWAPDeltaOrder,
   // Quote methods
   constructGetQuote,
   // different helpers
   constructGetPartnerFee,
-  constructGetBridgeInfo,
+  constructGetBridgeRoutes,
   constructIsTokenSupportedInDelta,
   constructEthersContractCaller, // same as constructEthersV5ContractCaller for backwards compatibility
   constructEthersV5ContractCaller,
@@ -482,37 +460,33 @@ export type {
   BuildNFTOrderInput,
   BuildNFTOrderDataInput,
   NFTOrdersUserParams,
-  //types for Delta methods
+  // types for Delta methods
   DeltaPrice,
-  BridgePrice,
   DeltaPriceParams,
+  DeltaPriceToken,
+  DeltaTokenAmount,
+  DeltaRoute,
+  DeltaRouteStep,
+  DeltaRouteBridge,
+  DeltaRouteBridgeContractParams,
+  BridgeTag,
+  BridgeRoute,
+  BuiltDeltaOrder,
   DeltaAuctionOrder,
   DeltaAuction,
-  DeltaAuctionDelta,
-  DeltaAuctionFillable,
-  DeltaAuctionTWAP,
-  DeltaAuctionTWAPBuy,
-  DeltaAuctionExternal,
-  DeltaAuctionProductive,
+  DeltaOrderStatus,
+  DeltaTokenSide,
+  DeltaTransaction,
   DeltaOrderUnion,
-  DeltaAuctionUnion,
   UnifiedDeltaOrderData,
-  DeltaAuctionStatus,
-  DeltaAuctionTransaction,
-  DeltaOrderFilterByStatus,
-  DeltaOrderFromAPI,
   CancelDeltaOrderData,
   SignableCancelDeltaOrderData,
-  // bridge part of DeltaOrder
-  BridgeMetadata,
-  BridgeStatus,
-  Bridge,
-  BridgeInfo,
-  AvailableBridge,
-  BridgeProtocolResponse,
-  BuildDeltaOrderDataParams,
-  BuildDeltaOrderFunctions,
   SignableDeltaOrderData,
+  // bridge part of DeltaOrder
+  Bridge,
+  BridgeProtocolResponse,
+  BuildDeltaOrderFunctions,
+  BuildDeltaOrderParams,
   DeltaOrderToPost,
   PostDeltaOrderFunctions,
   PostDeltaOrderParams,
@@ -521,8 +495,14 @@ export type {
   GetDeltaContractFunctions,
   GetDeltaPriceFunctions,
   GetDeltaOrdersFunctions,
+  GetBridgeRoutesFunctions,
+  GetAgentsListFunctions,
+  AgentList,
   ApproveTokenForDeltaFunctions,
   CancelDeltaOrderFunctions,
+  CancelDeltaOrder,
+  SignCancelDeltaOrderRequest,
+  PostCancelDeltaOrderRequest,
   DeltaTokenModuleFunctions,
   CancelAndWithdrawDeltaOrderParams,
   DepositNativeAndPreSignParams,
@@ -543,17 +523,15 @@ export type {
   SignableExternalOrderData,
   BuildExternalDeltaOrderParams,
   BuildExternalDeltaOrderFunctions,
-  SignExternalDeltaOrderFunctions,
   PostExternalDeltaOrderFunctions,
   PostExternalDeltaOrderParams,
   PreSignExternalDeltaOrderFunctions,
   // TWAP Delta types
   BuildTWAPDeltaOrderParams,
-  BuildTWAPSellOrderParams,
-  BuildTWAPBuyOrderParams,
+  BuildTWAPSellDeltaOrderParams,
+  BuildTWAPBuyDeltaOrderParams,
   BuildTWAPDeltaOrderFunctions,
   SignableTWAPOrderData,
-  SignTWAPDeltaOrderFunctions,
   PostTWAPDeltaOrderFunctions,
   PostTWAPDeltaOrderParams,
   PreSignTWAPDeltaOrderFunctions,
@@ -570,7 +548,6 @@ export type {
   ConstructProviderFetchInput,
   // other types
   GetPartnerFeeFunctions,
-  GetBridgeInfoFunctions,
   IsTokenSupportedInDeltaFunctions,
   Token,
   Address,
