+++ src/methods/delta/buildTWAPDeltaOrder.ts
@@ -1,187 +1,131 @@
+import { API_URL } from '../../constants';
 import type { ConstructFetchInput, RequestParameters } from '../../types';
-import { DEFAULT_BRIDGE } from './constants';
-import { constructGetDeltaContract } from './getDeltaContract';
-import type { BridgePrice } from './getDeltaPrice';
-import { constructGetPartnerFee } from './getPartnerFee';
-import {
-  buildTWAPSignableOrderData,
-  TWAPOrderCommonInput,
-  type BuildTWAPOrderDataInput,
-  type SignableTWAPOrderData,
-} from './helpers/buildTWAPOrderData';
-import { applySlippage, resolvePartnerFee } from './helpers/misc';
-import type { MarkOptional } from 'ts-essentials';
-export type { SignableTWAPOrderData } from './helpers/buildTWAPOrderData';
+import type { TWAPBuyDeltaOrder, TWAPDeltaOrder } from './helpers/types';
+import type { BuiltDeltaOrder, DeltaRoute } from './types';
+export type { BuiltDeltaOrder } from './types';
 
-type BuildTWAPDeltaOrderParamsBase = {
+type BuildTWAPDeltaOrderBase = {
   /** @description The address of the order owner */
   owner: string;
-  /** @description The address of the order beneficiary */
+  /** @description The address of the order beneficiary. Defaults to owner. */
   beneficiary?: string;
-  /** @description The address of the src token */
-  srcToken: string;
-  /** @description The address of the dest token. For Crosschain Order - destination token on the destination chain */
-  destToken: string;
-  /** @description The deadline for the order */
+  /** @description The deadline for the order (unix seconds) */
   deadline?: number;
-  /** @description The nonce of the order */
-  nonce?: number | string;
-  /** @description Optional permit signature for the src token */
+  /** @description The nonce of the order. Random if omitted. */
+  nonce?: string;
+  /** @description Optional permit signature for the src token. Defaults to "0x". */
   permit?: string;
-  /** @description Partner string */
+  /** @description Partner string. Passed to the server to resolve partner fee details. */
   partner?: string;
-  /** @description Destination Chain ID for Crosschain Orders */
-  destChainId?: number;
   /** @description Seconds between slice executions (min 60) */
   interval: number;
   /** @description Number of slices (min 2) */
   numSlices: number;
-  /** @description Slippage in basis points (bps). 10000 = 100%, 50 = 0.5% */
+  /** @description Slippage in basis points (bps). 10000 = 100%, 50 = 0.5%. Default 0. */
   slippage?: number;
-
-  /** @description price response received from /delta/prices (getDeltaPrice method) for a single slice */
-  deltaPrice: MarkOptional<
-    Pick<
-      BridgePrice,
-      | 'destAmount'
-      | 'partner'
-      | 'partnerFee'
-      | 'destToken'
-      | 'srcAmount'
-      | 'bridge'
-    >,
-    'partner' | 'partnerFee'
-  >;
-
-  /** @description partner fee in basis points (bps), 50bps=0.5% */
+  /** @description DeltaRoute from getDeltaPrice for a single slice */
+  route: DeltaRoute;
+  /** @description Partner fee in basis points (bps) */
   partnerFeeBps?: number;
-  /** @description partner address */
+  /** @description Partner address */
   partnerAddress?: string;
-  /** @description take surplus */
+  /** @description Take surplus flag */
   partnerTakesSurplus?: boolean;
-  /** @description A boolean indicating whether the surplus should be capped. True by default */
+  /** @description Whether the surplus should be capped. True by default. */
   capSurplus?: boolean;
   /** @description Metadata for the order, hex string */
   metadata?: string;
+  /** @description Designates the Order as partially fillable. Default false. */
+  partiallyFillable?: boolean;
+  /** @description If passed, the server will use this as SELL destAmount (as BUY srcAmount) and expectedAmount for each slice */
+  limitAmount?: string;
 };
 
-export type BuildTWAPSellOrderParams = BuildTWAPDeltaOrderParamsBase & {
-  /** @description Must be "TWAPOrder" for sell orders */
+export type BuildTWAPSellDeltaOrderParams = BuildTWAPDeltaOrderBase & {
   onChainOrderType: 'TWAPOrder';
-  /** @description Total source token amount across all slices */
+  /** @description Total source token amount across all slices. route.origin.input.amount must equal floor(totalSrcAmount / numSlices). */
   totalSrcAmount: string;
 };
 
-export type BuildTWAPBuyOrderParams = BuildTWAPDeltaOrderParamsBase & {
-  /** @description Must be "TWAPBuyOrder" for buy orders */
+export type BuildTWAPBuyDeltaOrderParams = BuildTWAPDeltaOrderBase & {
   onChainOrderType: 'TWAPBuyOrder';
-  /** @description Total destination token amount to buy across all slices */
+  /** @description Total destination token amount to buy across all slices. route.origin.output.amount must equal floor(totalDestAmount / numSlices). */
   totalDestAmount: string;
-  /** @description Maximum source token amount willing to spend */
+  /** @description Maximum source token amount willing to spend across all slices. */
   maxSrcAmount: string;
 };
 
 export type BuildTWAPDeltaOrderParams =
-  | BuildTWAPSellOrderParams
-  | BuildTWAPBuyOrderParams;
+  | BuildTWAPSellDeltaOrderParams
+  | BuildTWAPBuyDeltaOrderParams;
 
 type BuildTWAPDeltaOrder = (
   buildOrderParams: BuildTWAPDeltaOrderParams,
   requestParams?: RequestParameters
-) => Promise<SignableTWAPOrderData>;
+) => Promise<BuiltDeltaOrder<TWAPDeltaOrder | TWAPBuyDeltaOrder>>;
 
 export type BuildTWAPDeltaOrderFunctions = {
-  /** @description Build TWAP Orders (sell or buy) to be posted to Delta API for execution */
+  /** @description Build a Delta v2 TWAP Order (sell or buy) from a DeltaRoute via the server endpoint, ready to sign and post. */
   buildTWAPDeltaOrder: BuildTWAPDeltaOrder;
 };
 
 export const constructBuildTWAPDeltaOrder = (
   options: ConstructFetchInput
 ): BuildTWAPDeltaOrderFunctions => {
-  const { chainId } = options;
-
-  const { getDeltaContract } = constructGetDeltaContract(options);
-  const { getPartnerFee } = constructGetPartnerFee(options);
+  const { apiURL = API_URL, fetcher } = options;
+  const buildUrl = `${apiURL}/delta/v2/orders/build` as const;
 
   const buildTWAPDeltaOrder: BuildTWAPDeltaOrder = async (
     params,
     requestParams
   ) => {
-    const ParaswapDelta = await getDeltaContract(requestParams);
-    if (!ParaswapDelta) {
-      throw new Error(`Delta is not available on chain ${chainId}`);
-    }
-
-    const { partnerAddress, partnerFeeBps, partnerTakesSurplus } =
-      await resolvePartnerFee(params, getPartnerFee, requestParams);
-
-    const commonInput: TWAPOrderCommonInput = {
+    const commonBody = {
+      route: params.route,
       owner: params.owner,
       beneficiary: params.beneficiary,
-      srcToken: params.srcToken,
-      destToken: params.deltaPrice.destToken,
       deadline: params.deadline,
-      nonce: params.nonce?.toString(10),
+      nonce: params.nonce,
       permit: params.permit,
+      slippage: params.slippage,
+      limitAmount: params.limitAmount,
       metadata: params.metadata,
+      partiallyFillable: params.partiallyFillable,
+      partner: params.partner,
+      partnerAddress: params.partnerAddress,
+      partnerFeeBps: params.partnerFeeBps,
+      partnerTakesSurplus: params.partnerTakesSurplus,
+      capSurplus: params.capSurplus,
       interval: params.interval,
       numSlices: params.numSlices,
-      bridge: {
-        ...params.deltaPrice.bridge,
-        // TWAP child orders get their bridge data at execution time
-        protocolData: DEFAULT_BRIDGE.protocolData,
-      },
-
-      chainId,
-      paraswapDeltaAddress: ParaswapDelta,
-      partnerAddress,
-      partnerTakesSurplus,
-      partnerFeeBps,
-      capSurplus: params.capSurplus,
     };
 
-    let input: BuildTWAPOrderDataInput;
-
     if (params.onChainOrderType === 'TWAPOrder') {
-      const slippage = params.slippage ?? 0;
-      const destAmountPerSlice =
-        slippage > 0
-          ? applySlippage({
-              amount: params.deltaPrice.destAmount,
-              slippageBps: slippage,
-              increase: false,
-            })
-          : params.deltaPrice.destAmount;
-
-      input = {
-        ...commonInput,
-        onChainOrderType: 'TWAPOrder',
-        destAmountPerSlice,
-        totalSrcAmount: params.totalSrcAmount,
-      };
-    } else {
-      const slippage = params.slippage ?? 0;
-      const maxSrcAmount =
-        slippage > 0
-          ? applySlippage({
-              amount: params.maxSrcAmount,
-              slippageBps: slippage,
-              increase: true,
-            })
-          : params.maxSrcAmount;
-
-      input = {
-        ...commonInput,
-        onChainOrderType: 'TWAPBuyOrder',
-        totalDestAmount: params.totalDestAmount,
-        maxSrcAmount,
-      };
+      return fetcher<BuiltDeltaOrder<TWAPDeltaOrder>>({
+        url: buildUrl,
+        method: 'POST',
+        data: {
+          ...commonBody,
+          side: 'SELL',
+          orderType: 'TWAPOrder',
+          totalSrcAmount: params.totalSrcAmount,
+        },
+        requestParams,
+      });
     }
 
-    return buildTWAPSignableOrderData(input);
+    return fetcher<BuiltDeltaOrder<TWAPBuyDeltaOrder>>({
+      url: buildUrl,
+      method: 'POST',
+      data: {
+        ...commonBody,
+        side: 'BUY',
+        orderType: 'TWAPBuyOrder',
+        totalDestAmount: params.totalDestAmount,
+        maxSrcAmount: params.maxSrcAmount,
+      },
+      requestParams,
+    });
   };
 
-  return {
-    buildTWAPDeltaOrder,
-  };
+  return { buildTWAPDeltaOrder };
 };
