+++ src/methods/delta/getAgentsList.ts
