+++ src/methods/delta/getDeltaOrders.ts
@@ -3,25 +3,11 @@ import { constructSearchString } from '../../helpers/misc';
 import type {
   Address,
   ConstructFetchInput,
+  PaginatedResponse,
   RequestParameters,
 } from '../../types';
-import type {
-  DeltaAuction,
-  DeltaAuctionStatus,
-  DeltaOrderType,
-  OnChainOrderMap,
-} from './helpers/types';
-
-/** @deprecated Use DeltaAuction directly */
-export type DeltaOrderFromAPI = DeltaAuction;
-
-export type DeltaOrderFilterByStatus =
-  | DeltaAuctionStatus
-  | 'INSUFFICIENT_BALANCE'
-  | 'INSUFFICIENT_ALLOWANCE'
-  | 'INVALIDATED'
-  | 'ACTIVE'
-  | 'INACTIVE';
+import type { DeltaOrderType, OnChainOrderType } from './helpers/types';
+import type { DeltaOrderStatus, DeltaAuction } from './types';
 
 type GetDeltaOrderById = (
   orderId: string,
@@ -33,44 +19,32 @@ type GetDeltaOrderByHash = (
   requestParams?: RequestParameters
 ) => Promise<DeltaAuction>;
 
-type OrdersFilter<T extends keyof OnChainOrderMap = keyof OnChainOrderMap> = {
-  /** @description Order.owner to fetch Delta Order for */
+type OrdersFilter = {
+  /** @description `order.owner` to fetch Delta Orders for. */
   userAddress: Address;
-  /** @description Pagination option, page. Default 1 */
+  /** @description Pagination option. Default 1. */
   page?: number;
-  /** @description Pagination option, limit. Default 100 */
+  /** @description Pagination option. Default 100, max 1000. */
   limit?: number;
-  /** @description Filter by chainId, without this filter, orders from all chains are returned */
+  /** @description Filter by chainId. Omitted = orders across all chains. */
   chainId?: number[];
-  /**
-   * @description
-   * Filter by any known DeltaAuctionStatus and some custom statuses:
-   * - **INSUFFICIENT_BALANCE** —  returned as SUSPENDED from API
-   * - **INSUFFICIENT_ALLOWANCE** —  returned as SUSPENDED from API
-   * - **INVALIDATED** —  returned as FAILED from API
-   * - **ACTIVE** —  All orders with NOT_STARTED, RUNNING, EXECUTING, CANCELLING or SUSPENDED statuses.
-   * - **INACTIVE** —  All orders with EXECUTED, FAILED, EXPIRED, CANCELLED or INVALIDATED statuses.
-   */
-  status?: DeltaOrderFilterByStatus[];
-  /** @description Filter by type. MARKET, LIMIT. Orders with both types are returned if not specified */
+  /** @description Filter by integrator-facing status. */
+  status?: DeltaOrderStatus[];
+  /** @description Filter by order type. MARKET or LIMIT. */
   type?: DeltaOrderType;
-  /** @description Filter by on-chain order type. Order, ExternalOrder. Orders of all types are returned if not specified */
-  onChainOrderType?: T;
+  /** @description Filter by on-chain order type. */
+  onChainOrderType?: OnChainOrderType;
 };
+
 type OrderFiltersQuery = Omit<OrdersFilter, 'chainId' | 'status'> & {
   chainId?: string;
   status?: string;
 };
 
-type GetDeltaOrders = {
-  <T extends keyof OnChainOrderMap>(
-    options: OrdersFilter<T> & { onChainOrderType: T },
-    requestParams?: RequestParameters
-  ): Promise<DeltaAuction<T>[]>;
-  (options: OrdersFilter, requestParams?: RequestParameters): Promise<
-    DeltaAuction[]
-  >;
-};
+type GetDeltaOrders = (
+  options: OrdersFilter,
+  requestParams?: RequestParameters
+) => Promise<PaginatedResponse<DeltaAuction>>;
 
 type GetRequiredBalanceParams = {
   userAddress: Address;
@@ -80,13 +54,17 @@ type GetRequiredBalanceParams = {
 type GetRequiredBalance = (
   userParams: GetRequiredBalanceParams,
   requestParams?: RequestParameters
-) => Promise<Record<string, string>>; // token -> balance in Limit Orders
+) => Promise<Record<string, string>>; // token -> required balance across open Delta orders
 
 export type GetDeltaOrdersFunctions = {
+  /** @description Fetch a single order by its UUID. */
   getDeltaOrderById: GetDeltaOrderById;
+  /** @description Fetch a single order by its EIP-712 order hash. */
   getDeltaOrderByHash: GetDeltaOrderByHash;
+  /** @description List Delta orders with the v2 pagination envelope. */
   getDeltaOrders: GetDeltaOrders;
-  getRequiredBalanceForDeltaLimitOrders: GetRequiredBalance;
+  /** @description Required balance per token across the user's open Delta v2 orders. Pass `tokenAddress` to narrow the result to a single token. */
+  getRequiredBalanceForDeltaOrders: GetRequiredBalance;
 };
 
 export const constructGetDeltaOrders = ({
@@ -94,21 +72,18 @@ export const constructGetDeltaOrders = ({
   fetcher,
   chainId,
 }: ConstructFetchInput): GetDeltaOrdersFunctions => {
-  const baseUrl = `${apiURL}/delta/orders` as const;
+  const baseUrl = `${apiURL}/delta/v2/orders` as const;
 
   const getDeltaOrderById: GetDeltaOrderById = async (
     orderId,
     requestParams
   ) => {
     const fetchURL = `${baseUrl}/${orderId}` as const;
-
-    const order = await fetcher<DeltaAuction>({
+    return fetcher<DeltaAuction>({
       url: fetchURL,
       method: 'GET',
       requestParams,
     });
-
-    return order;
   };
 
   const getDeltaOrderByHash: GetDeltaOrderByHash = async (
@@ -116,22 +91,14 @@ export const constructGetDeltaOrders = ({
     requestParams
   ) => {
     const fetchURL = `${baseUrl}/hash/${orderHash}` as const;
-
-    const order = await fetcher<DeltaAuction>({
+    return fetcher<DeltaAuction>({
       url: fetchURL,
       method: 'GET',
       requestParams,
     });
-
-    return order;
   };
 
-  const getDeltaOrders: GetDeltaOrders = async <
-    T extends keyof OnChainOrderMap = keyof OnChainOrderMap
-  >(
-    options: OrdersFilter<T>,
-    requestParams?: RequestParameters
-  ) => {
+  const getDeltaOrders: GetDeltaOrders = async (options, requestParams) => {
     const chainIdString = options.chainId?.join(',');
     const statusString = options.status?.join(',');
 
@@ -147,16 +114,14 @@ export const constructGetDeltaOrders = ({
 
     const fetchURL = `${baseUrl}${search}` as const;
 
-    const orders = await fetcher<DeltaAuction<T>[]>({
+    return fetcher<PaginatedResponse<DeltaAuction>>({
       url: fetchURL,
       method: 'GET',
       requestParams,
     });
-
-    return orders;
   };
 
-  const getRequiredBalanceForDeltaLimitOrders: GetRequiredBalance = async (
+  const getRequiredBalanceForDeltaOrders: GetRequiredBalance = async (
     userParams,
     requestParams
   ) => {
@@ -179,6 +144,6 @@ export const constructGetDeltaOrders = ({
     getDeltaOrderById,
     getDeltaOrderByHash,
     getDeltaOrders,
-    getRequiredBalanceForDeltaLimitOrders,
+    getRequiredBalanceForDeltaOrders,
   };
 };
