+++ src/sdk/full.ts
@@ -13,10 +13,6 @@ import {
   DeltaOrderHandlers,
 } from '../methods/delta';
 import {
-  constructAllDeltaOrdersHandlers as constructAllDeltaV2OrdersHandlers,
-  DeltaOrderHandlers as DeltaV2OrderHandlers,
-} from '../methods/deltaV2';
-import {
   constructGetQuote,
   GetQuoteFunctions,
 } from '../methods/quote/getQuote';
@@ -30,7 +26,6 @@ export type AllSDKMethods<TxResponse> = {
   /** @deprecated NFT Orders are deprecated and will be removed in a future version. */
   nftOrders: NFTOrderHandlers<TxResponse>;
   delta: DeltaOrderHandlers<TxResponse>;
-  deltaV2: DeltaV2OrderHandlers<TxResponse>;
   quote: GetQuoteFunctions;
 } & Required<ConstructBaseInput>;
 
@@ -46,8 +41,6 @@ export const constructFullSDK = <TxResponse = any>(
     constructAllNFTOrdersHandlers(config);
   const delta: DeltaOrderHandlers<TxResponse> =
     constructAllDeltaOrdersHandlers(config);
-  const deltaV2: DeltaV2OrderHandlers<TxResponse> =
-    constructAllDeltaV2OrdersHandlers(config);
   const quote = constructGetQuote(config);
 
   return {
@@ -55,7 +48,6 @@ export const constructFullSDK = <TxResponse = any>(
     limitOrders,
     nftOrders,
     delta,
-    deltaV2,
     quote,
     apiURL: config.apiURL ?? API_URL,
     chainId: config.chainId,
